+++ HISTORY.md
@@ -1,6 +1,12 @@
 # Changelog
+
+# 0.6.2
+* FIX preprocessing with VST floats and inf
+* FIX plotly display
+
 # 0.6.1
 * FIX data loading
+
 # 0.6.0
 * ADD mzTAB support
 * ENH color Volcano Plot data points using list of protein names `color_list=your_protein_list`